+++ lib/live_motion/motion.ex
@@ -3,7 +3,7 @@ defmodule LiveMotion.Motion do
   Represents configuration for LiveMotion animations.
   """
 
-  defstruct keyframes: %{}, transition: nil, exit: nil
+  defstruct keyframes: %{}, transition: nil, exit: nil, defer: false
 
   defimpl Phoenix.HTML.Safe, for: LiveMotion.Motion do
     def to_iodata(%LiveMotion.Motion{} = motion) do