+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.1.2] (2022-04-14)
+
+### Added
+
+- Added `LiveMotion.JS.show/1` to allow animations when showing elements.
+- Added `defer` option to `LiveMotion.motion` to prevent animations on mount.
+
 ## [v0.1.1] (2022-04-12)
 
 ### Fixed